Safari-609.1.11.1/LayoutTests/fast/text/find-diacritical-mark-expected.txt

Tests find for strings with letters that have diacritical marks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFind('cafe', 'cafÉ', forward, CaseInsensitive) is "not found"
PASS testFind('cafe', 'cafÉ', forward, CaseSensitive) is "not found"
PASS testFind('cafe', 'café', forward, CaseInsensitive) is "not found"
PASS testFind('cafe', 'café', forward, CaseSensitive) is "not found"
PASS testFind('cafÉ', 'cafE', forward, CaseInsensitive) is "0, 4"
PASS testFind('cafÉ', 'cafE', forward, CaseSensitive) is "not found"
PASS testFind('cafÉ', 'cafe', forward, CaseInsensitive) is "0, 4"
PASS testFind('cafÉ', 'cafe', forward, CaseSensitive) is "not found"
PASS testFind('cafÉ', 'café', forward, CaseInsensitive) is "0, 4"
PASS testFind('cafÉ', 'café', forward, CaseSensitive) is "not found"
PASS testFind('café', 'cafE', forward, CaseInsensitive) is "0, 4"
PASS testFind('café', 'cafE', forward, CaseSensitive) is "not found"
PASS testFind('café', 'cafe', forward, CaseInsensitive) is "0, 4"
PASS testFind('café', 'cafe', forward, CaseSensitive) is "not found"
PASS testFind('café', 'cafÉ', forward, CaseInsensitive) is "0, 4"
PASS testFind('café', 'cafÉ', forward, CaseSensitive) is "not found"
PASS testFind('café', 'café', forward, CaseInsensitive) is "0, 4"
PASS testFind('café', 'café', forward, CaseSensitive) is "0, 4"

Thai letters:
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ทั้ง', forward, CaseInsensitive) is "49, 53"
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ทั้ง', forward, CaseSensitive) is testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ทั้ง', forward, CaseInsensitive)
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ร่วม', forward, CaseInsensitive) is "not found"
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ร่วม', forward, CaseSensitive) is testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ร่วม', forward, CaseInsensitive)
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ดู', forward, CaseInsensitive) is "not found"
PASS testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ดู', forward, CaseSensitive) is testFind('สื่ออิเล็กทรอนิกส์ อินเทอร์เน็ต หรือทางใดทางหนึ่งทั้งสิ้น', 'ดู', forward, CaseInsensitive)
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ทั้ง', forward, CaseInsensitive) is "not found"
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ทั้ง', forward, CaseSensitive) is testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ทั้ง', forward, CaseInsensitive)
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ร่วม', forward, CaseInsensitive) is "42, 46"
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ร่วม', forward, CaseSensitive) is testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ร่วม', forward, CaseInsensitive)
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ดู', forward, CaseInsensitive) is "not found"
PASS testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ดู', forward, CaseSensitive) is testFind('หากท่านใดประสงค์จะพิมพ์แจกเป็นธรรมทาน หรือร่วมสนับสนุนสื่อธรรมะโปรดติดต่อ', 'ดู', forward, CaseInsensitive)
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ทั้ง', forward, CaseInsensitive) is "not found"
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ทั้ง', forward, CaseSensitive) is testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ทั้ง', forward, CaseInsensitive)
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ร่วม', forward, CaseInsensitive) is "not found"
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ร่วม', forward, CaseSensitive) is testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ร่วม', forward, CaseInsensitive)
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ดู', forward, CaseInsensitive) is "55, 57"
PASS testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ดู', forward, CaseSensitive) is testFind('และเทคโนโลยี ก็พัฒนามาจนทำให้ผู้สนใจสามารถฟังไฟล์เสียง ดูวีดีโอ ได้จากอินเทอร์เน็ต ตามเวลา', 'ดู', forward, CaseInsensitive)
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ทั้ง', forward, CaseInsensitive) is "not found"
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ทั้ง', forward, CaseSensitive) is testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ทั้ง', forward, CaseInsensitive)
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ร่วม', forward, CaseInsensitive) is "27, 31"
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ร่วม', forward, CaseSensitive) is testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ร่วม', forward, CaseInsensitive)
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ดู', forward, CaseInsensitive) is "not found"
PASS testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ดู', forward, CaseSensitive) is testFind('และขอขอบคุณทุกท่านที่มีส่วนร่วมในการทำให้งานนี้สำเร็จจนมาอยู่ในมือท่านในขณะนี้', 'ดู', forward, CaseInsensitive)
PASS successfullyParsed is true

TEST COMPLETE
